+++ octobot_channels/producer.py
@@ -104,6 +104,14 @@ class Producer:
         """
         pass
 
+    async def wait_for_processing(self) -> None:
+        """
+        Should be used only with SupervisedConsumers
+        It will wait until all consumers have notified that their consume() method have ended
+        :return: None
+        """
+        await asyncio.gather(*[consumer.queue.join() for consumer in self.channel.get_consumers()])
+
     async def stop(self):
         """
         Stops non-triggered tasks management