+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2019-09-07
+### Added
+- Supervised Consumer that notify the consumption end
+
+### Fixed
+- Consumer tests
+
 ## [1.2.0] - 2019-09-04
 ### Added
 - Channel add_new_consumer method to add a new consumer with filters